+++ src/type.d.ts
@@ -2,6 +2,7 @@ interface ITask extends Record<string,any> {
     id?: string
     title: string
     description?: string
+    category?: string
     complete?: boolean
 }